remove cookie pair with Max-Age=0
It seems to be common to use the setcookie function to set the cookie value to null during the process of deleting a cookie. PHP will set the value to "deleted" in this case. But we shouldn’t just check this value, the correct method is checking the expires value or the max-age value.
Since the script runs for a very short time, we only need to check for Max-Age=0.